Fixing server_test
