+++ ext/storage/git_state_manager_test.go
@@ -14,58 +14,13 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func gitPrepare(t *testing.T, s *sous.State, remotepath, outpath string) {
-
-	clobberDir(t, remotepath)
-	clobberDir(t, outpath)
-
-	runCmd(t, remotepath, "git", "init", "--template=/dev/null", "--bare")
-
-	remoteAbs, err := filepath.Abs(remotepath)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	runCmd(t, outpath, "git", "init", "--template=/dev/null")
-	runCmd(t, outpath, "git", "config", "user.email", "sous-test@testing.example.com")
-	runCmd(t, outpath, "git", "config", "user.name", "sous-test@testing.example.com")
-	runCmd(t, outpath, "git", "remote", "add", "origin", "file://"+remoteAbs)
-
-	dsm := NewDiskStateManager(outpath)
-	dsm.WriteState(s, testUser)
-
-	runCmd(t, outpath, "git", "add", ".")
-	runCmd(t, outpath, "git", "commit", "--no-gpg-sign", "-a", "-m", "birthday")
-	runCmd(t, outpath, "git", "push", "-u", "origin", "master")
-}
-
-func clobberDir(t *testing.T, path string) {
-	if err := os.RemoveAll(path); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.MkdirAll(path, os.ModePerm); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func runCmd(t *testing.T, path string, cmd ...string) {
-	gitCmd := exec.Command(cmd[0], cmd[1:]...)
-	gitCmd.Dir = path
-	out, err := gitCmd.CombinedOutput()
-	if err != nil {
-		t.Fatalf("%q errored: %v\n %s", strings.Join(cmd, " "), err, out)
-	}
-}
-
-var testUser = sous.User{Name: "Test User", Email: "test@user.com"}
-
 func TestGitStateManager_WriteState_success(t *testing.T) {
 	require := require.New(t)
 
 	s := exampleState()
 
 	clobberDir(t, "testdata/result")
-	gitPrepare(t, s, "testdata/remote", "testdata/out")
+	PrepareTestGitRepo(t, s, "testdata/remote", "testdata/out")
 
 	gsm := NewGitStateManager(NewDiskStateManager("testdata/out"))
 
@@ -91,9 +46,7 @@ func TestGitStateManager_WriteState_success(t *testing.T) {
 func TestGitStateManager_WriteState_multiple_manifests(t *testing.T) {
 
 	s := exampleState()
-
-	gitPrepare(t, s, "testdata/remote", "testdata/out")
-
+	PrepareTestGitRepo(t, s, "testdata/remote", "testdata/out")
 	gsm := NewGitStateManager(NewDiskStateManager("testdata/out"))
 
 	// Modify one of the manifests.
@@ -127,7 +80,7 @@ func TestGitReadState(t *testing.T) {
 	require := require.New(t)
 
 	s := exampleState()
-	gitPrepare(t, s, "testdata/remote", "testdata/out")
+	PrepareTestGitRepo(t, s, "testdata/remote", "testdata/out")
 	gsm := NewGitStateManager(NewDiskStateManager("testdata/out"))
 
 	actual, err := gsm.ReadState()
